+++ MESManager/MESManager.Infrastructure/Migrations/20260203082709_AddFestiviTable.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace MESManager.Infrastructure.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddFestiviTable : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+
+        }
+    }
+}